[#9] Simplified GUI widget usage
- Container no longer accepts lambas (lambdas should not be used instead of a proper widget)
- VectorText widget created to replace the lambdas used in main.cpp for displaying player vectors
- Some changes to dynamic and rotary to get pointers to player position and rotation
- Renamed UI.h and UI.cpp to GUI.h and GUI.cpp to match the class name
- Some indent changes are applied but should applied to all files in future commit
