+++ Voxal Engine.vcxproj
@@ -104,14 +104,14 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>D:\glfw-3.3.8.bin.WIN64\include;D:\VulkanSDK\1.3.261.1\Include;D:\glm;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>D:\Dev\VulkanSDK\1.3.250.0\Include;D:\Dev\glfw-3.3.8.bin.WIN64\include;D:\Dev\glm;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <LanguageStandard_C>Default</LanguageStandard_C>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalLibraryDirectories>D:\glfw-3.3.8.bin.WIN64\lib-vc2022;D:\VulkanSDK\1.3.261.1\Lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalLibraryDirectories>D:\Dev\glfw-3.3.8.bin.WIN64\lib-vc2022;D:\Dev\VulkanSDK\1.3.250.0\Lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
       <AdditionalDependencies>vulkan-1.lib;glfw3.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
@@ -133,6 +133,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <None Include="README.md" />
+    <None Include="src\VectorText.h" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\Buffer.h" />
@@ -162,7 +163,7 @@
     <ClInclude Include="src\Rasterizer.h" />
     <ClInclude Include="src\Rotary.h" />
     <ClInclude Include="src\SwapChain.h" />
-    <ClInclude Include="src\UI.h" />
+    <ClInclude Include="src\GUI.h" />
     <ClInclude Include="src\Updatable.h" />
     <ClInclude Include="src\volk.h" />
     <ClInclude Include="src\Voxel.h" />
@@ -175,7 +176,6 @@
     <ClCompile Include="src\Camera.cpp" />
     <ClCompile Include="src\Chunk.cpp" />
     <ClCompile Include="src\Config.cpp" />
-    <ClCompile Include="src\Container.cpp" />
     <ClCompile Include="src\Descriptor.cpp" />
     <ClCompile Include="src\Device.cpp" />
     <ClCompile Include="src\Observer.cpp" />
@@ -197,7 +197,7 @@
     <ClCompile Include="src\SwapChain.cpp" />
     <ClCompile Include="src\Updatable.cpp" />
     <ClCompile Include="src\Utils.hpp" />
-    <ClCompile Include="src\UI.cpp" />
+    <ClCompile Include="src\GUI.cpp" />
     <ClCompile Include="src\volk.c" />
     <ClCompile Include="src\Window.cpp" />
     <ClCompile Include="src\World.cpp" />
