+++ src/Container.h
@@ -1,21 +1,29 @@
 #pragma once
-#include <functional>
 #include <string>
-#include <variant>
 #include <vector>
 
+#include "imgui.h"
 #include "Widget.h"
 
 namespace gm {
-	class Container: public Widget {
-		std::string title;
-		std::vector<std::variant<Widget*, std::function<void()>>> widgets{};
+  class Container: public Widget {
+    std::string title;
+    std::vector<Widget*> widgets{};
 
-	public:
-		Container(const std::string& title, std::vector<std::variant<Widget*, std::function<void()>>> widgets): title{title}, widgets{std::move(widgets)} {}
+  public:
+    Container(const std::string& title, std::vector<Widget*> widgets): title{title}, widgets{std::move(widgets)} {}
 
-		void addWidget(Widget* w) { widgets.emplace_back(w); }
-		void addWidget(std::function<void()> f) { widgets.emplace_back(f); }
-		void render() override;
-	};
+    ~Container() {
+      for(const auto& w : widgets) { delete w; }
+      widgets.clear();
+    }
+
+    void addWidget(Widget* w) { widgets.emplace_back(w); }
+
+    void render() override {
+      ImGui::Begin(title.c_str());
+      for(auto w : widgets) { w->render(); }
+      ImGui::End();
+    }
+  };
 }
