+++ src/Graphic.cpp
@@ -4,12 +4,11 @@
 #include "Utils.hpp"
 #include "Window.h"
 #include "Device.h"
-#include "UI.h"
 #include "imgui_impl_vulkan.h"
 
 static VKAPI_ATTR VkBool32 VKAPI_CALL debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity, VkDebugUtilsMessageTypeFlagsEXT messageType, const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData, void* pUserData) {
-	std::cerr << pCallbackData->pMessage << "\n" << std::endl;
-	return VK_FALSE;
+  std::cerr << pCallbackData->pMessage << "\n" << std::endl;
+  return VK_FALSE;
 }
 
 static VkDebugUtilsMessengerCreateInfoEXT debugCreateInfo = {.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT, .messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT, .messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT, .pfnUserCallback = debugCallback};
@@ -21,73 +20,69 @@ bool enableValidationLayers = true;
 #endif
 
 namespace gm {
-	Module::Module() {
-		addExtension(Window::getRequiredExtensions());
-		createVkInstance();
+  Module::Module() {
+    addExtension(Window::getRequiredExtensions());
+    createVkInstance();
 
-		window = std::make_unique<Window>(instance);
-		Device::Initializer deviceInit{instance};
-		deviceInit.setTargetSurface(window->getSurface());
-		//deviceInit.addExtension();
-		device = deviceInit.init();
-		window->init(device.get());
-	}
+    window = std::make_unique<Window>(instance, "Window");
+    Device::Initializer deviceInit{instance};
+    deviceInit.setTargetSurface(window->getSurface());
+    //deviceInit.addExtension();
+    device = deviceInit.init();
+    window->init(device.get());
+  }
 
-	Module::~Module() {
-		vkDeviceWaitIdle(device->getVkDevice());
-	}
+  Module::~Module() { vkDeviceWaitIdle(device->getVkDevice()); }
 
-	void Module::createVkInstance() {
-		volkInitialize();
-		if(enableValidationLayers) {
-			//Check for validation layer support
-			uint32_t layerCount;
-			vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
-			std::vector<VkLayerProperties> availableLayers(layerCount);
-			vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
-			for(const char* layerName: validationLayers) {
-				bool layerFound = false;
+  void Module::createVkInstance() {
+    volkInitialize();
+    if(enableValidationLayers) {
+      //Check for validation layer support
+      uint32_t layerCount;
+      vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
+      std::vector<VkLayerProperties> availableLayers(layerCount);
+      vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
+      for(const char* layerName : validationLayers) {
+        bool layerFound = false;
 
-				for(const auto& layerProperties: availableLayers) {
-					if(strcmp(layerName, layerProperties.layerName) == 0) {
-						layerFound = true;
-						break;
-					}
-				}
+        for(const auto& layerProperties : availableLayers) {
+          if(strcmp(layerName, layerProperties.layerName) == 0) {
+            layerFound = true;
+            break;
+          }
+        }
 
-				if(!layerFound) {
-					Utils::warn("Validation layers requested, but not available!");
-					enableValidationLayers = false;
-					break;
-				}
-			}
-			instanceExtensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-		}
+        if(!layerFound) {
+          Utils::warn("Validation layers requested, but not available!");
+          enableValidationLayers = false;
+          break;
+        }
+      }
+      instanceExtensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
+    }
 
-		VkApplicationInfo appInfo = {.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO, .pEngineName = "Voxal Engine", .engineVersion = VK_MAKE_VERSION(0, 2, 0), .apiVersion = VK_API_VERSION_1_3};
+    VkApplicationInfo appInfo = {.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO, .pEngineName = "Voxal Engine", .engineVersion = VK_MAKE_VERSION(0, 2, 0), .apiVersion = VK_API_VERSION_1_3};
 
-		VkInstanceCreateInfo createInfo = {.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO, .pApplicationInfo = &appInfo, .enabledExtensionCount = static_cast<uint32_t>(instanceExtensions.size()), .ppEnabledExtensionNames = instanceExtensions.data(),};
-		if(enableValidationLayers) {
-			createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
-			createInfo.ppEnabledLayerNames = validationLayers.data();
-			createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
-		} else {
-			createInfo.enabledLayerCount = 0;
-			createInfo.pNext = nullptr;
-		}
-		VK_CHECK_RESULT(vkCreateInstance(&createInfo, nullptr, &instance), "Failed to create Vulkan instance!")
-		volkLoadInstance(instance);
-		ImGui_ImplVulkan_LoadFunctions(
-			[](const char* function_name, void* vulkan_instance) {
-				return vkGetInstanceProcAddr(*(reinterpret_cast<VkInstance*>(vulkan_instance)), function_name);
-			}, &instance
-		);
+    VkInstanceCreateInfo createInfo = {.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO, .pApplicationInfo = &appInfo, .enabledExtensionCount = static_cast<uint32_t>(instanceExtensions.size()), .ppEnabledExtensionNames = instanceExtensions.data(),};
+    if(enableValidationLayers) {
+      createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
+      createInfo.ppEnabledLayerNames = validationLayers.data();
+      createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
+    } else {
+      createInfo.enabledLayerCount = 0;
+      createInfo.pNext = nullptr;
+    }
+    VK_CHECK_RESULT(vkCreateInstance(&createInfo, nullptr, &instance), "Failed to create Vulkan instance!")
+    volkLoadInstance(instance);
+    ImGui_ImplVulkan_LoadFunctions(
+    [](const char* function_name, void* vulkan_instance) { return vkGetInstanceProcAddr(*(reinterpret_cast<VkInstance*>(vulkan_instance)), function_name); }, &instance
+    );
 
-		if(enableValidationLayers) VK_CHECK_RESULT(vkCreateDebugUtilsMessengerEXT(instance, &debugCreateInfo, nullptr, &debugMessenger), "Failed to Create Debug Messenger!")
-	}
+    if(enableValidationLayers) VK_CHECK_RESULT(vkCreateDebugUtilsMessengerEXT(instance, &debugCreateInfo, nullptr, &debugMessenger), "Failed to Create Debug Messenger!")
+  }
 
-	Module& Module::getInstance() {
-		static Module singleton;
-		return singleton;
-	}
+  Module& Module::getInstance() {
+    static Module singleton;
+    return singleton;
+  }
 }
