+++ src/Player.h
@@ -6,15 +6,15 @@
 #include "Rotary.h"
 
 class Player: public Observer, public obj::Base, public obj::Dynamic, public obj::Rotary {
-	float speed = 7; //TODO add attribute system
+  float speed = 7; //TODO add attribute system
 public:
-	Player(obj::Transform transform, World* world);
+  Player(obj::Transform transform, World* world);
 
-	void move(glm::vec3 direction) override;
-	void setPosition(glm::vec3 pos) override;
-	glm::vec3 getPosition() const override { return transform.position; }
+  void move(glm::vec3 direction) override;
+  void setPosition(glm::vec3 pos) override;
+  glm::vec3& getPosition() override { return transform.position; }
 
-	void rotate(glm::vec3 rotation) override;
-	void setRotation(glm::vec3 rotation) override;
-	glm::vec3 getRotation() override { return transform.rotation; }
+  void rotate(glm::vec3 rotation) override;
+  void setRotation(glm::vec3 rotation) override;
+  glm::vec3& getRotation() override { return transform.rotation; }
 };
